+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add a new CLI component, `lint-dependency-groups`, which can be used to lint
+  dependency groups.
+
 ## 0.1.1
 
 - Fix a bug in cycle detection for nontrivial cycles